CSS + Bug fixes
preview css fixes

youtube css fixes

icon css fixes

rem useless import
