+++ src/common/components/Emote.module.css
@@ -1,0 +1,9 @@
+.emoteImage {
+  width: 28px;
+  height: 28px;
+  object-fit: contain;
+}
+
+.placeholder {
+  background-color: var(--bttv-rs-bg-card);
+}
