+++ src/modules/emote_menu/components/Emotes.module.css
@@ -50,6 +50,7 @@
   display: flex;
   align-items: center;
   width: 100%;
+  height: 300px;
   flex-direction: column;
   row-gap: 8px;
   justify-content: center;
