+++ src/modules/emote_menu/components/Preview.jsx
@@ -3,8 +3,8 @@ import {Icon} from '@rsuite/icons';
 import IconButton from 'rsuite/IconButton';
 import * as faStar from '@fortawesome/free-solid-svg-icons/faStar';
 import styles from './Preview.module.css';
-import {createSrcSet} from '../../../utils/image.js';
 import FontAwesomeSvgIcon from '../../../common/components/FontAwesomeSvgIcon.jsx';
+import Emote from '../../../common/components/Emote.jsx';
 
 export default function Preview({emote, isFavorite}) {
   if (emote == null) return null;
@@ -12,14 +12,7 @@ export default function Preview({emote, isFavorite}) {
   return (
     <div className={styles.preview} key={emote.code}>
       <div className={styles.content}>
-        <div className={styles.emote}>
-          <img
-            alt={emote.name}
-            className={styles.emoteImage}
-            srcSet={createSrcSet(emote.images, ['2x', '4x'])}
-            src={emote.images['2x'] || emote.images['1x']}
-          />
-        </div>
+        <Emote emote={emote} />
         <div>
           <div className={styles.emoteCode}>{emote.code}</div>
           <div>from {emote.category.displayName}</div>
