+++ src/modules/emote_menu/components/Preview.module.css
@@ -1,4 +1,5 @@
 .preview {
+  box-sizing: border-box;
   display: flex;
   justify-content: space-between;
   padding: 8px;
@@ -15,8 +16,8 @@
 }
 
 .emoteImage {
-  width: 36px;
-  height: 36px;
+  width: 28px;
+  height: 28px;
   object-fit: contain;
 }
